Wrote getAll function in cars-model. Wrote first GET endpoint in cars-router. Linked cars-router to server. Tests 1-2 are passing

+++ data/migrations/01-make_cars_table.js
@@ -5,7 +5,7 @@ exports.up = async function (knex) {
     table.string('vin').unique().notNullable()
     table.string('make').notNullable()
     table.string('model').notNullable()
-    table.decimal('mileage').notNullable()
+    table.integer('mileage').notNullable()
     table.string('title')
     table.string('transmission')
   })
